Move default soul config under src/config
