+++ dimp-server/src/config/default_guild_soul.txt
