+++ dimp-server/src/config/guild-soul.ts
@@ -2,7 +2,7 @@ import { readFileSync } from "node:fs"
 import { fileURLToPath } from "node:url"
 
 const defaultGuildSoulPath = fileURLToPath(
-    new URL("./default_soul.txt", import.meta.url)
+    new URL("./default_guild_soul.txt", import.meta.url)
 )
 
 export const DEFAULT_GUILD_SOUL = readFileSync(
